pyproject.toml
[project]
name = "api-template-dev"
version = "0.1.0"
description = "FastAPI service template for development and testing"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    # Core dependencies for the template code
    "authlib>=1.6.4",
    "fastapi>=0.116.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "starlette>=0.47.3",
    "uvicorn[standard]>=0.35.0",
    "httpx>=0.27.2",
    "cachetools>=5.3.3",
    "fastapi-limiter>=0.1.6",
    "aioredis>=2.0.0",
    "structlog>=23.0.0",
    "loguru>=0.7.3",
    # Template development dependencies
    "cookiecutter>=2.1.1",
    "cruft>=2.15.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-authlib>=1.6.2.20250914",
]

[project.scripts]
init-db = "generator.runtime.init_db:init_db"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["F401", "F811"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q --tb=short"
python_files = "tests.py test_*.py *_tests.py"
pythonpath = ["."]
testpaths = ["tests"]